* Rename DoubleColumnTransfer to Transfer * Added a condition, so that in context creator one cannot proceed to map configuration, unless enabled plugins contain at least one bodyPlugin, to prevent plugins being undefined in PluginsContainer
